Updated all the files except LICENSE, and README.md to be in the develop folder/Updated shapes.test.js to include the triangle, circle, and square for testing, and then also created a test suite for the triangle class however I have not installed any npm packages so I have not tested it yet/Updated shapes.js to include classes for the triangle, circle, and square shapes so that I can call them in the shapes.test.js file for testing however I have not yet tested them because I dont have the npm packages installed/Updated the .gitignore file to remove the node_modules/ from the top on line 1 so that im not going against the DRY principal/Updated index.js to include pseudo code and a skeleton for the const questions that will be prompted using the npm i inquirer package
Updated all the files except LICENSE, and README.md to be in the develop folder so I can generate have better readability for the file structure. Updated shapes.test.js to include the triangle, circle, and square for testing, and then also created a test suite for the triangle class however I have not installed any npm packages so I have not tested it yet but I will install those soon and fix. Updated shapes.js to include classes for the triangle, circle, and square shapes so that I can call them in the shapes.test.js file for testing however I have not yet tested them because I dont have the npm packages installed however will install the jest and inquirer soon. Updated the .gitignore file to remove the node_modules/ from the top on line 1 so that im not going against the DRY principal. Updated index.js to include pseudo code and a skeleton for the const questions that will be prompted using the npm i inquirer package however its not installed yet.