Update index.js so that there is questions for a prompt when the application is run so that the user can pick what they want to have on their logo, and added a write to file function that tells the user that the data is saved if the last prompted question has yes selected else the console will log "No data to save." when the last question in the prompt is answered with the selection no/Updated shapes.js so that there is a generateSvgLogo function for an init function that will take in the generateSvgLogo data that will pass through the write to file function in order for the write to file function to actually generate the data that is recieved from the prompted questions however im not sure if it works yet still have to test out the triangle.
Update index.js so that there is questions for a prompt when the application is run so that the user can pick what they want to have on their logo, and added a write to file function that tells the user that the data is saved if the last prompted question has yes selected else the console will log "No data to save." when the last question in the prompt is answered with the selection no this way the user knows that the logo has been saved else not saved. Updated shapes.js so that there is a generateSvgLogo function for an init function that will take in the generateSvgLogo data that will pass through the write to file function in order for the write to file function to actually generate the data that is recieved from the prompted questions however im not sure if it works yet still have to test out the triangle to make sure that the triangle will return with a color when the code is ran.