Update .gitignore to include package-lock.json so that when pushed to repo the file is ignored/Updated file structure so there is no longer a develop folder, added a logo.html, added a cli.js to the lib, added a svg.js to the lib/Updated the logo.html to add a boiler plate for the logo.html.
Update .gitignore to include package-lock.json so that when pushed to repo the file is ignored. Updated file structure so there is no longer a develop folder, added a logo.html, added a cli.js to the lib, added a svg.js to the lib so I can seperate for concerns. Updated the logo.html to add a boiler plate for the logo.html so that there is place for the logo to go however if else wont go to the examples folder yet will fix.