Updated the index.js to include the init function that uses the inquirer library to get questions prompted in the cli so that when the user answers the questions the then method is used to handle the resolved promise and execute the callback function with the answers, I can run the file using node index.js from the CLI and have the application initialize from within the callback function so that the generateSvgLogo is called with the answers as an argument to generate SVG logo data, and the result of the function is stored in a variable data, after that the data generated is logged into the console using console.log(data) finally the writeToFile function called will write the data as an argument to the write the generated data to a file when it is run/Updated the cli.js to include the required files using variables and the require method, took the questions array and shortened it, added a function to the text question for the logo using the validate method to check if the length of the value is strictly equal to three then have it return true, if else return with a prompt that asks the user to Please enter exactly three characters, included the writeToFile function, and changed it so that it writes the file inside of the example folder, included the init function to initialize the prompt questions when the index.js (application) file is ran from the CLI/Updated the shapes.js to include just one class of shape, and then used the super method to initialize the shapeColor that is inherited from the shape class, then each sub class has a render method that returns an svg string representing the shape with the specified color, finally this files code exports the triangle, circle, and square classes so that they can be apart of other parts in the codebase/Updated the svg.js to include the shape sub classes Triangle, circle, and square, then the code for svg.js has a function called generateSvgLogo that will take in the data object as a parameter and the object is expected to have the properties of shape, shapecolor, textcolor, and text, then inside of the function it will check the value of data to determine which shape to instantiate, after the shape instantiates the function generates an svg code that will include the rendered shape using the render method of the shape instance, it also includes text color, and the text for the logo so with that the function construct will then an HTML code with a template literal string that will be returned as a string, lastly the generatesSvgLogo function is modularized for other parts of the code so it can be reused.
Updated the index.js to include the init function that uses the inquirer library to get questions prompted in the cli so that when the user answers the questions the then method is used to handle the resolved promise and execute the callback function with the answers, I can run the file using node index.js from the CLI and have the application initialize from within the callback function so that the generateSvgLogo is called with the answers as an argument to generate SVG logo data, and the result of the function is stored in a variable data, after that the data generated is logged into the console using console.log(data) finally the writeToFile function called will write the data as an argument to the write the generated data to a file when it is run this way the user gets the logo they want! Updated the cli.js to include the required files using variables and the require method, took the questions array and shortened it, added a function to the text question for the logo using the validate method to check if the length of the value is strictly equal to three then have it return true, if else return with a prompt that asks the user to Please enter exactly three characters, after that I included the writeToFile function, and changed it so that it writes the file inside of the example folder, included the init function to initialize the prompt questions when the index.js (application) file is ran from the CLI using node.js. Updated the shapes.js to include just one class of shape, and then used the super method to initialize the shapeColor that is inherited from the shape class, then made each sub class use a render method that returns an svg string representing the shape with the specified color, finally this files code exports the triangle, circle, and square classes so that they can be apart of other parts in the codebase. Updated the svg.js to include the shape sub classes Triangle, Circle, and Square, then the code for svg.js has a function called generateSvgLogo that will take in the data object as a parameter and the object is expected to have the properties of shape, shapecolor, textcolor, and text, then inside of the function it will check the value of data to determine which shape to instantiate, after the shape instantiates the function generates an svg code that will include the rendered shape using the render method of the shape instance, it also include text color, and the text for the logo so with that the function will then construct an HTML code with a template literal string that will be returned as a string, lastly the generatesSvgLogo function is modularized for other parts of the code so it can be reused.