Updated svg.js to include some styling for displaying the logo in the template literal string/Updated logo.html to include styling for the generated logo so that its centered.
Updated svg.js to include some styling for displaying the logo in the template literal string however it does not work will fix. Updated logo.html to include styling for the generated logo so that its centered however it does not load up the logo.svg in the html yet will fix.