Updated index.js to include comments describing the functionalities of the code/Updated cli.js to include comments describing the functionalities of the code/Updated shapes.js to include comments describing the functionalities of the code/ Updated shapes.test.js to test if a new triangle instance will accept blue as a color, however the test failed as the expected result was blue and the recieved was undefined, will make the test pass for each shape/ Updated svg.js to include comments describing the functionalities of the code/ Updated logo.html to remove a commented out piece of code that was for implementing the generated logo into an html file and having it pop up in the browser however that does not work yet will fix.
Updated index.js to include comments describing the functionalities of the code so its easier to read. Updated cli.js to include comments describing the functionalities of the code so its easier to read. Updated shapes.js to include comments describing the functionalities of the code so its easier to read. Updated shapes.test.js to test if a new triangle instance will accept blue as a color, however the test failed as the expected result was blue and the recieved was undefined, will make the test pass for each shape so that each shape is tested, and will fix. Updated svg.js to include comments describing the functionalities of the code so its easier to read. Updated logo.html to remove a commented out piece of code that was for implementing the generated logo into an html file and having it pop up in the browser however that does not work yet will fix.