Update removed logo.html because its not needed/Updated README.md file with relevant screenshots, an example video, and a link to google docs for a sample of the svg file/Updated the examples folder with images of generated shapes for the README.md/Updated the shapes.js to change the size of the circle/Updated shapes.test.js to include working tests for the shape sub-classes/Updated the svg.js to include just a return on the svg code to remove the html code because it is not needed.
Update removed logo.html because its not needed for this. Updated README.md file with relevant screenshots, an example video, and a link to google docs for a sample of the svg file so that I can show the user how to use the application. Updated the examples folder with images of generated shapes for the README.md so the user can see screenshots of some generated shapes. Updated the shapes.js to change the size of the circle so that its not a half circle I changed the r sizing. Updated shapes.test.js to include working tests for the shape sub-classes with a describe shape, then describe the sub class, then it should have the color of blue when a new instance is made then pass the it statement into the function to describe the expectation of having a color blue, having a shape relevant to whichever sub class is described, and then expecting it to be a this shapeColor which should turn out to be blue, and then repeated that for the other shapes and the tests worked after I got them to fail. Updated the svg.js to include just a return on the svg code to remove the html code because it is not needed for the code to display the logo.svg.